requirements.txt
sqlfluff
faker
psycopg2-binary
sqlalchemy
requests
